Move commit lines transformer into the commits package
