+++ apps/desktop/src/components/BaseBranch.svelte
@@ -4,7 +4,7 @@
 	import CommitCard from '$components/CommitCard.svelte';
 	import InfoMessage from '$components/InfoMessage.svelte';
 	import { BaseBranchService } from '$lib/baseBranch/baseBranchService';
-	import { transformAnyCommit } from '$lib/commitLines/transformers';
+	import { transformAnyCommit } from '$lib/commits/transformers';
 	import { getForge } from '$lib/forge/interface/forge';
 	import { ModeService } from '$lib/modes/service';
 	import { groupByCondition } from '$lib/utils/array';
